checkpoint: renames "Attempt/Error/NonActionableError.ts" to "Attempt/Error/NonActionable.ts"
