+++ src/library/Attempt/Error/Creation.ts
@@ -1,4 +1,4 @@
-import Attempt_Error_NonActionable from './NonActionableError';
+import Attempt_Error_NonActionable from './NonActionable';
 
 class Attempt_Error_Creation
   extends Attempt_Error_NonActionable {
