+++ src/library/Attempt/Error/NonActionable.ts
