+++ src/library/Attempt/Error/assertions/assertPotentiallyActionable.ts
@@ -2,7 +2,7 @@ import Attempt_Error_Creation from '../Creation';
 
 import {
   default as Attempt_Error_NonActionable,
-} from '../NonActionableError';
+} from '../NonActionable';
 
 import type {
   PotentiallyActionable,
